requirements.txt
hydra-core
pydantic
torch
torchvision
pytorch-lightning
torchmetrics
omegaConf